refactor: extrair componente Filter da Home
