+++ src/components/Filter/styles.module.scss
@@ -1,0 +1,60 @@
+.form-group {
+    margin: 1rem 0;
+    position: relative;
+    display: flex;
+    padding: 1rem 1.5rem;
+    align-items: center;
+    border-radius: var(--radius);
+    background: var(--elevation-2);
+    box-shadow: 0px 0px 4px 0px rgba(0, 0, 0, 0.07), 0px 3px 4px 0px rgba(0, 0, 0, 0.07);
+
+    &.dropdown {
+        > div:first-child {
+            display: flex;
+            align-items: center;
+            cursor: pointer;
+        }
+
+        svg {
+            margin-left: 1.2rem;
+            transform: rotateZ(180deg);
+            transition: all 0.2s ease;
+        }
+
+        .dropdown-wrapper {
+            position: absolute;
+            left: 0;
+            top: 64px;
+            display: inline-flex;
+            width: 100%;
+            flex-direction: column;
+            border-radius: var(--radius);
+            background: var(--elevation-2);
+            box-shadow: 0px 0px 4px 0px rgba(0, 0, 0, 0.07), 0px 3px 4px 0px rgba(0, 0, 0, 0.07);
+            max-height: 0;
+            overflow: hidden;
+            transition: all 0.2s ease;
+
+            a {
+                cursor: pointer;
+                padding: 0.75rem 1.5rem;
+                opacity: 0.75;
+
+                &:hover {
+                    opacity: 1;
+                }
+            }
+        }
+
+        &.open {
+            svg {
+                transform: rotateZ(0);
+            }
+
+            .dropdown-wrapper {
+                max-height: 300px;
+                padding: 0.75rem 0;
+            }
+        }
+    }
+}
