+++ src/pages/Home/index.jsx
@@ -1,17 +1,14 @@
-import { useEffect, useRef } from 'react'
+import { useEffect } from 'react'
 import styles from './styles.module.scss'
 import { useState } from 'react'
-import { LuChevronUp } from 'react-icons/lu'
 import { CountryCard } from '../../components/CountryCard'
 import { SearchBar } from '../../components/SearchBar'
+import { Filter } from '../../components/Filter'
 
 const URL = 'https://restcountries.com/v3.1/'
 
-const Regions = ['Africa', 'America', 'Asia', 'Europe', 'Oceania']
-
 export default function Home() {
     const [countries, setCountries] = useState([])
-    const dropdownRef = useRef(null)
     const fetchCountries = async ({ name = null, region = null }) => {
         let filterBy = name ? `name/${name}` : 'all'
         filterBy = region ? `region/${region}` : filterBy
@@ -25,31 +22,11 @@ export default function Home() {
         fetchCountries({})
     }, [])
 
-    const handleFilterByRegion = region => {
-        fetchCountries({ region })
-        handleDropdown()
-    }
-
-    const handleDropdown = () => {
-        dropdownRef.current.classList.toggle(styles.open)
-    }
     return (
         <div className={`container`}>
             <div className={styles.row}>
                 <SearchBar />
-                <div className={`${styles['form-group']} ${styles.dropdown}`} ref={dropdownRef}>
-                    <div onClick={handleDropdown}>
-                        Filter by Region <LuChevronUp />
-                    </div>
-                    <div className={styles['dropdown-wrapper']}>
-                        <a onClick={() => handleFilterByRegion(null)}>All Regions</a>
-                        {Regions.map(region => (
-                            <a key={region} onClick={() => handleFilterByRegion(region)}>
-                                {region}
-                            </a>
-                        ))}
-                    </div>
-                </div>
+                <Filter />
             </div>
             <div className={styles['countries-grid']}>
                 {countries?.length &&
